Add files via upload
This Excel VBA macro file demonstrates how to use the Sleep function in order to time the use of your code. In this case when to save the workbook. Feel free to clone this repo and try to enhance it or download it to your machine. Feel free to use the code in your own projects. Put a reference, to my repo, in your code comments when doing that.